Finish README
This text (in the field "Commit changes") describes the changes that I did above, when I was revising the masterversion in the branch, that is,when I wrote the changes that will be saved as a commit. (Basically, the Commit change is, describing what I am doing when I am writing a branch in GitHub). By the way, the title of "Commit changes" that now reads "Finish README", was initially Updates to README. I changed it into Finish README because it is what is used in the example.